add back button, scroll view, fix detailscreen, connect with detailscreen and homescreen

+++ app/src/main/java/com/example/fooddelivery/PopularData.kt
@@ -11,7 +11,7 @@ data class PopularData(
     val price: Double,
     val rate: Double,
     val description: String,
-    val calorie: Double,
-    val scheduleTime: Double,
+    val calorie: Int,
+    val scheduleTime: Int,
     val ingredients: List<Int>
 ): Parcelable
